requirements.txt
pandas==2.2.2
psycopg2-binary==2.9.9
flask==3.0.3
python-dotenv==1.0.1
